+++ src/db/group.js
@@ -1,6 +1,7 @@
 const database = require('../database')
 
 const DuplicateGroupException = class extends Error {}
+const UserAlreadyInGroupException = class extends Error {}
 
 const Group = class {
   constructor (id) {
@@ -59,7 +60,20 @@ const Group = class {
   }
 
   addUser (userId) {
-    require('./user_in_group').add_User(userId, this._id)
+    // check if the user is already in the group
+    const result = database.query('SELECT * FROM User_in_Group WHERE user_id = :value AND group_id = :group_id', {
+      value: userId,
+      group_id: this._id
+    })
+
+    if (result.length !== 0) {
+      throw new UserAlreadyInGroupException()
+    }
+
+    database.query('INSERT INTO User_in_Group (user_id, group_id) VALUES (:user_id, :group_id)', {
+      user_id: userId,
+      group_id: this._id
+    })
   }
 
   _select (property) {
@@ -123,5 +137,6 @@ module.exports = {
   by_name: byName,
   create: create,
   get_all: getAllGroups,
-  DuplicateGroupException: DuplicateGroupException
+  DuplicateGroupException: DuplicateGroupException,
+  UserAlreadyInGroupException: UserAlreadyInGroupException
 }
